requirements.txt
asgiref==3.7.2
blinker==1.8.1
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.6
Django==4.2.2
dnspython==2.6.1
dsinternals==1.2.4
Flask==3.0.3
future==1.0.0
impacket==0.11.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
ldap3==2.9.1
ldapdomaindump==0.9.4
MarkupSafe==2.1.5
packaging==24.0
pillow==10.3.0
pluggy==1.4.0
protobuf==4.21.12
pyasn1==0.6.0
pycparser==2.22
pycryptodomex==3.20.0
pyOpenSSL==24.1.0
pytest==8.1.1
six==1.16.0
smbls==2.0.0
sqlparse==0.4.4
Werkzeug==3.0.2
